move to core where possible
