+++ packages/grafana-ui/src/components/index.ts
@@ -291,11 +291,6 @@ export * from './uPlot/geometries';
 export * from './uPlot/plugins';
 export { type PlotTooltipInterpolator, type PlotSelection } from './uPlot/types';
 export { type UPlotConfigPrepFn } from './uPlot/config/UPlotConfigBuilder';
-export { GraphNG, type GraphNGProps, FIXED_UNIT } from './GraphNG/GraphNG';
-export { TimeSeries } from './TimeSeries/TimeSeries';
-export { useGraphNGContext } from './GraphNG/hooks';
-export { preparePlotFrame, buildScaleKey } from './GraphNG/utils';
-export { type GraphNGLegendEvent } from './GraphNG/types';
 export * from './PanelChrome/types';
 export { Label as BrowserLabel } from './BrowserLabel/Label';
 export { PanelContainer } from './PanelContainer/PanelContainer';
