+++ public/app/core/components/GraphNG/hooks.ts
