+++ public/app/core/components/GraphNG/utils.ts
@@ -1,15 +1,8 @@
-import { DataFrame, Field, FieldConfig, FieldType, outerJoinDataFrames, TimeRange } from '@grafana/data';
-import {
-  AxisPlacement,
-  GraphDrawStyle,
-  GraphFieldConfig,
-  ScaleDistribution,
-  ScaleDistributionConfig,
-} from '@grafana/schema';
+import { DataFrame, Field, FieldType, outerJoinDataFrames, TimeRange } from '@grafana/data';
+import { applyNullInsertThreshold } from '@grafana/data/src/transformations/transformers/nulls/nullInsertThreshold';
+import { nullToUndefThreshold } from '@grafana/data/src/transformations/transformers/nulls/nullToUndefThreshold';
+import { GraphDrawStyle } from '@grafana/schema';
 
-import { FIXED_UNIT } from './GraphNG';
-import { applyNullInsertThreshold } from './nullInsertThreshold';
-import { nullToUndefThreshold } from './nullToUndefThreshold';
 import { XYFieldMatchers } from './types';
 
 function isVisibleBarField(f: Field) {
@@ -144,37 +137,4 @@ export function preparePlotFrame(frames: DataFrame[], dimFields: XYFieldMatchers
   }
 
   return null;
-}
-
-export function buildScaleKey(config: FieldConfig<GraphFieldConfig>, fieldType: FieldType) {
-  const defaultPart = 'na';
-
-  const scaleRange = `${config.min !== undefined ? config.min : defaultPart}-${
-    config.max !== undefined ? config.max : defaultPart
-  }`;
-
-  const scaleSoftRange = `${config.custom?.axisSoftMin !== undefined ? config.custom.axisSoftMin : defaultPart}-${
-    config.custom?.axisSoftMax !== undefined ? config.custom.axisSoftMax : defaultPart
-  }`;
-
-  const scalePlacement = `${
-    config.custom?.axisPlacement !== undefined ? config.custom?.axisPlacement : AxisPlacement.Auto
-  }`;
-
-  const scaleUnit = config.unit ?? FIXED_UNIT;
-
-  const scaleDistribution = config.custom?.scaleDistribution
-    ? getScaleDistributionPart(config.custom.scaleDistribution)
-    : ScaleDistribution.Linear;
-
-  const scaleLabel = Boolean(config.custom?.axisLabel) ? config.custom!.axisLabel : defaultPart;
-
-  return `${scaleUnit}/${scaleRange}/${scaleSoftRange}/${scalePlacement}/${scaleDistribution}/${scaleLabel}/${fieldType}`;
-}
-
-function getScaleDistributionPart(config: ScaleDistributionConfig) {
-  if (config.type === ScaleDistribution.Log) {
-    return `${config.type}${config.log}`;
-  }
-  return config.type;
 }
