+++ public/app/plugins/panel/timeseries/plugins/ThresholdControlsPlugin.tsx
@@ -2,7 +2,8 @@ import React, { useState, useLayoutEffect, useMemo, useRef } from 'react';
 import uPlot from 'uplot';
 
 import { FieldConfigSource, ThresholdsConfig, getValueFormat, FieldType } from '@grafana/data';
-import { UPlotConfigBuilder, buildScaleKey } from '@grafana/ui';
+import { UPlotConfigBuilder } from '@grafana/ui';
+import { buildScaleKey } from '@grafana/ui/src/components/uPlot/internal';
 
 import { ThresholdDragHandle } from './ThresholdDragHandle';
