+++ public/app/plugins/panel/xychart/dims.ts
@@ -1,5 +1,5 @@
 import { DataFrame, Field, FieldMatcher, FieldType, getFieldDisplayName } from '@grafana/data';
-import { XYFieldMatchers } from '@grafana/ui/src/components/GraphNG/types';
+import { XYFieldMatchers } from 'app/core/components/GraphNG/types';
 
 import { XYDimensionConfig } from './panelcfg.gen';
